+++ src/Phinx/Migration/Manager/Environment.php
@@ -32,6 +32,7 @@ use Phinx\Db\Adapter\AdapterFactory;
 use Phinx\Db\Adapter\AdapterInterface;
 use Phinx\Migration\MigrationInterface;
 use Phinx\Seed\SeedInterface;
+use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
 class Environment
@@ -45,6 +46,11 @@ class Environment
      * @var array
      */
     protected $options;
+
+    /**
+     * @var InputInterface
+     */
+    protected $input;
 
     /**
      * @var OutputInterface
@@ -197,6 +203,28 @@ class Environment
     }
 
     /**
+     * Sets the console input.
+     *
+     * @param InputInterface $input
+     * @return Environment
+     */
+    public function setInput(InputInterface $input)
+    {
+        $this->input = $input;
+        return $this;
+    }
+
+    /**
+     * Gets the console input.
+     *
+     * @return InputInterface
+     */
+    public function getInput()
+    {
+        return $this->input;
+    }
+
+    /**
      * Sets the console output.
      *
      * @param OutputInterface $output Output
@@ -312,6 +340,10 @@ class Environment
                 ->getWrapper($this->options['wrapper'], $adapter);
         }
 
+        if ($this->getInput()) {
+            $adapter->setInput($this->getInput());
+        }
+        
         if ($this->getOutput()) {
             $adapter->setOutput($this->getOutput());
         }